refactor: factor out EvalError into eval_error.rs
Since interpreter.rs is getting very long.
